allow the use of JSON configs in .jshintrc
This allows you to used the same .jshintrc for the node version of
jshint and the jshint.vim plugin, which needs a comment style config.

jshint.vim checks whether the .jshintrc file starts with a `/*jshint`
directive. If it doesn't, it assumes that the .jshintrc file is a JSON
object. It passes the contents of the .jshintrc to transform.js (based
on runjslint.js) This parses the JSON and builds up a comment string
that is necessary.